+++ packages/dart/lib/src/hub_adapter.dart
@@ -11,6 +11,7 @@ import 'scope.dart';
 import 'sentry.dart';
 import 'sentry_client.dart';
 import 'sentry_options.dart';
+import 'telemetry/metric/metric.dart';
 import 'tracing.dart';
 
 /// Hub adapter to make Integrations testable
@@ -199,6 +200,10 @@ class HubAdapter implements Hub {
 
   @override
   FutureOr<void> captureLog(SentryLog log) => Sentry.currentHub.captureLog(log);
+
+  @override
+  Future<void> captureMetric(SentryMetric metric) =>
+      Sentry.currentHub.captureMetric(metric);
 
   @override
   void setAttributes(Map<String, SentryAttribute> attributes) =>