+++ packages/dart/test/sentry_test.dart
@@ -8,6 +8,7 @@ import 'package:sentry/src/dart_exception_type_identifier.dart';
 import 'package:sentry/src/event_processor/deduplication_event_processor.dart';
 import 'package:sentry/src/logs_enricher_integration.dart';
 import 'package:sentry/src/feature_flags_integration.dart';
+import 'package:sentry/src/telemetry/metric/metrics_setup_integration.dart';
 import 'package:sentry/src/telemetry/processing/processor_integration.dart';
 import 'package:test/test.dart';
 
@@ -268,6 +269,27 @@ void main() {
       expect(integration.callCalls, 1);
     });
 
+    test('should add $MetricsSetupIntegration', () async {
+      late SentryOptions optionsReference;
+      final options = defaultTestOptions();
+
+      await Sentry.init(
+        options: options,
+        (options) {
+          options.dsn = fakeDsn;
+          optionsReference = options;
+        },
+        appRunner: appRunner,
+      );
+
+      expect(
+        optionsReference.integrations
+            .whereType<MetricsSetupIntegration>()
+            .length,
+        1,
+      );
+    });
+
     test('should add default integrations', () async {
       late SentryOptions optionsReference;
       final options = defaultTestOptions();
@@ -355,7 +377,7 @@ void main() {
       );
     }, onPlatform: {'vm': Skip()});
 
-    test('should add feature flag FeatureFlagsIntegration', () async {
+    test('should add feature flag $FeatureFlagsIntegration', () async {
       await Sentry.init(
         options: defaultTestOptions(),
         (options) => options.dsn = fakeDsn,